+++ full-moon/tests/roblox_cases/fail/parser/generic_declare_packs_last/source.lua
@@ -1,0 +1,1 @@
+type Bar<T..., U> = nil