+++ full-moon/tests/roblox_cases/fail/parser/generic_must_declare_default/source.lua
@@ -1,0 +1,1 @@
+type Foo<X, Y, Z = string, P> = nil